moved msg_create_trade_route
